+++ src/index.js
@@ -37,6 +37,7 @@ export {default as reverseIterate} from './reverseIterate.js';
 export {default as setIndex} from './setIndex.js';
 export {default as splice} from './splice.js';
 export {default as split} from './split.js';
+export {default as splitFirst} from './splitFirst.js';
 export {default as splitIndex} from './splitIndex.js';
 export {default as splitLast} from './splitLast.js';
 export {default as toArray} from './toArray.js';