+++ types/index.d.ts
@@ -33,6 +33,7 @@ export {default as reverseIterate} from './reverseIterate';
 export {default as setIndex} from './setIndex';
 export {default as splice} from './splice';
 export {default as split} from './split';
+export {default as splitFirst} from './splitFirst';
 export {default as splitIndex} from './splitIndex';
 export {default as splitLast} from './splitLast';
 export {default as toArray} from './toArray';